+++ DefaultResources/Decal.shader
@@ -14,7 +14,7 @@ CGPROGRAM
 
 sampler2D _MainTex;
 sampler2D _DecalTex;
-float4 _Color;
+fixed4 _Color;
 
 struct Input {
 	float2 uv_MainTex;
@@ -22,7 +22,7 @@ struct Input {
 };
 
 void surf (Input IN, inout SurfaceOutput o) {
-	half4 c = tex2D(_MainTex, IN.uv_MainTex);
+	fixed4 c = tex2D(_MainTex, IN.uv_MainTex);
 	half4 decal = tex2D(_DecalTex, IN.uv_DecalTex);
 	c.rgb = lerp (c.rgb, decal.rgb, decal.a);
 	c *= _Color;
