+++ DefaultResources/Illumin-Bumped.shader
@@ -16,7 +16,7 @@ CGPROGRAM
 sampler2D _MainTex;
 sampler2D _BumpMap;
 sampler2D _Illum;
-float4 _Color;
+fixed4 _Color;
 
 struct Input {
 	float2 uv_MainTex;
@@ -25,8 +25,8 @@ struct Input {
 };
 
 void surf (Input IN, inout SurfaceOutput o) {
-	half4 tex = tex2D(_MainTex, IN.uv_MainTex);
-	half4 c = tex * _Color;
+	fixed4 tex = tex2D(_MainTex, IN.uv_MainTex);
+	fixed4 c = tex * _Color;
 	o.Albedo = c.rgb;
 	o.Emission = c.rgb * tex2D(_Illum, IN.uv_Illum).a;
 	o.Alpha = c.a;
