+++ DefaultResources/Illumin-Parallax.shader
@@ -19,7 +19,7 @@ sampler2D _MainTex;
 sampler2D _BumpMap;
 sampler2D _ParallaxMap;
 sampler2D _Illum;
-float4 _Color;
+fixed4 _Color;
 float _Parallax;
 
 struct Input {
@@ -36,7 +36,7 @@ void surf (Input IN, inout SurfaceOutput o) {
 	IN.uv_BumpMap += offset;
 	IN.uv_Illum += offset;
 	
-	half4 c = tex2D(_MainTex, IN.uv_MainTex) * _Color;
+	fixed4 c = tex2D(_MainTex, IN.uv_MainTex) * _Color;
 	o.Albedo = c.rgb;
 	o.Emission = c.rgb * tex2D(_Illum, IN.uv_Illum).a;
 	o.Alpha = c.a;
