+++ DefaultResources/Internal-GUITextureClipText.shader
@@ -23,13 +23,13 @@ Shader "Hidden/Internal-GUITextureClipText"
 
 			struct appdata_t {
 				float4 vertex : POSITION;
-				float4 color : COLOR;
+				fixed4 color : COLOR;
 				float2 texcoord : TEXCOORD0;
 			};
 
 			struct v2f {
 				float4 vertex : POSITION;
-				float4 color : COLOR;
+				fixed4 color : COLOR;
 				float2 texcoord : TEXCOORD0;
 				float2 texgencoord : TEXCOORD1;
 			};
@@ -38,7 +38,7 @@ Shader "Hidden/Internal-GUITextureClipText"
 			sampler2D _GUIClipTexture;
 
 			uniform float4 _MainTex_ST;
-			uniform float4 _Color;
+			uniform fixed4 _Color;
 			uniform float4x4 _GUIClipTextureMatrix;
 			
 			v2f vert (appdata_t v)
@@ -52,9 +52,9 @@ Shader "Hidden/Internal-GUITextureClipText"
 				return o;
 			}
 
-			half4 frag (v2f i) : COLOR
+			fixed4 frag (v2f i) : COLOR
 			{
-				float4 col = _Color;
+				fixed4 col = _Color;
 				col.a *= tex2D(_MainTex, i.texcoord).a * tex2D(_GUIClipTexture, i.texgencoord).a;
 				return col;
 			}
