+++ DefaultResources/Normal-BumpSpec.shader
@@ -15,8 +15,8 @@ CGPROGRAM
 
 sampler2D _MainTex;
 sampler2D _BumpMap;
-float4 _Color;
-float _Shininess;
+fixed4 _Color;
+half _Shininess;
 
 struct Input {
 	float2 uv_MainTex;
@@ -24,7 +24,7 @@ struct Input {
 };
 
 void surf (Input IN, inout SurfaceOutput o) {
-	half4 tex = tex2D(_MainTex, IN.uv_MainTex);
+	fixed4 tex = tex2D(_MainTex, IN.uv_MainTex);
 	o.Albedo = tex.rgb * _Color.rgb;
 	o.Gloss = tex.a;
 	o.Alpha = tex.a * _Color.a;
