+++ DefaultResources/Normal-Diffuse.shader
@@ -11,14 +11,14 @@ CGPROGRAM
 #pragma surface surf Lambert
 
 sampler2D _MainTex;
-float4 _Color;
+fixed4 _Color;
 
 struct Input {
 	float2 uv_MainTex;
 };
 
 void surf (Input IN, inout SurfaceOutput o) {
-	half4 c = tex2D(_MainTex, IN.uv_MainTex) * _Color;
+	fixed4 c = tex2D(_MainTex, IN.uv_MainTex) * _Color;
 	o.Albedo = c.rgb;
 	o.Alpha = c.a;
 }
