+++ DefaultResources/Normal-Parallax.shader
@@ -17,7 +17,7 @@ CGPROGRAM
 sampler2D _MainTex;
 sampler2D _BumpMap;
 sampler2D _ParallaxMap;
-float4 _Color;
+fixed4 _Color;
 float _Parallax;
 
 struct Input {
@@ -32,7 +32,7 @@ void surf (Input IN, inout SurfaceOutput o) {
 	IN.uv_MainTex += offset;
 	IN.uv_BumpMap += offset;
 	
-	half4 c = tex2D(_MainTex, IN.uv_MainTex) * _Color;
+	fixed4 c = tex2D(_MainTex, IN.uv_MainTex) * _Color;
 	o.Albedo = c.rgb;
 	o.Alpha = c.a;
 	o.Normal = UnpackNormal(tex2D(_BumpMap, IN.uv_BumpMap));
