+++ DefaultResources/Reflect-BumpSpec.shader
@@ -20,9 +20,9 @@ sampler2D _MainTex;
 sampler2D _BumpMap;
 samplerCUBE _Cube;
 
-float4 _Color;
-float4 _ReflectColor;
-float _Shininess;
+fixed4 _Color;
+fixed4 _ReflectColor;
+half _Shininess;
 
 struct Input {
 	float2 uv_MainTex;
@@ -32,8 +32,8 @@ struct Input {
 };
 
 void surf (Input IN, inout SurfaceOutput o) {
-	half4 tex = tex2D(_MainTex, IN.uv_MainTex);
-	half4 c = tex * _Color;
+	fixed4 tex = tex2D(_MainTex, IN.uv_MainTex);
+	fixed4 c = tex * _Color;
 	o.Albedo = c.rgb;
 	
 	o.Gloss = tex.a;
@@ -42,7 +42,7 @@ void surf (Input IN, inout SurfaceOutput o) {
 	o.Normal = UnpackNormal(tex2D(_BumpMap, IN.uv_BumpMap));
 	
 	float3 worldRefl = WorldReflectionVector (IN, o.Normal);
-	half4 reflcol = texCUBE (_Cube, worldRefl);
+	fixed4 reflcol = texCUBE (_Cube, worldRefl);
 	reflcol *= tex.a;
 	o.Emission = reflcol.rgb * _ReflectColor.rgb;
 	o.Alpha = reflcol.a * _ReflectColor.a;
