+++ DefaultResources/Reflect-Bumped.shader
@@ -18,8 +18,8 @@ sampler2D _MainTex;
 sampler2D _BumpMap;
 samplerCUBE _Cube;
 
-float4 _Color;
-float4 _ReflectColor;
+fixed4 _Color;
+fixed4 _ReflectColor;
 
 struct Input {
 	float2 uv_MainTex;
@@ -29,14 +29,14 @@ struct Input {
 };
 
 void surf (Input IN, inout SurfaceOutput o) {
-	half4 tex = tex2D(_MainTex, IN.uv_MainTex);
-	half4 c = tex * _Color;
+	fixed4 tex = tex2D(_MainTex, IN.uv_MainTex);
+	fixed4 c = tex * _Color;
 	o.Albedo = c.rgb;
 	
 	o.Normal = UnpackNormal(tex2D(_BumpMap, IN.uv_BumpMap));
 	
 	float3 worldRefl = WorldReflectionVector (IN, o.Normal);
-	half4 reflcol = texCUBE (_Cube, worldRefl);
+	fixed4 reflcol = texCUBE (_Cube, worldRefl);
 	reflcol *= tex.a;
 	o.Emission = reflcol.rgb * _ReflectColor.rgb;
 	o.Alpha = reflcol.a * _ReflectColor.a;
