+++ DefaultResources/Reflect-Diffuse.shader
@@ -15,8 +15,8 @@ CGPROGRAM
 sampler2D _MainTex;
 samplerCUBE _Cube;
 
-float4 _Color;
-float4 _ReflectColor;
+fixed4 _Color;
+fixed4 _ReflectColor;
 
 struct Input {
 	float2 uv_MainTex;
@@ -24,11 +24,11 @@ struct Input {
 };
 
 void surf (Input IN, inout SurfaceOutput o) {
-	half4 tex = tex2D(_MainTex, IN.uv_MainTex);
-	half4 c = tex * _Color;
+	fixed4 tex = tex2D(_MainTex, IN.uv_MainTex);
+	fixed4 c = tex * _Color;
 	o.Albedo = c.rgb;
 	
-	half4 reflcol = texCUBE (_Cube, IN.worldRefl);
+	fixed4 reflcol = texCUBE (_Cube, IN.worldRefl);
 	reflcol *= tex.a;
 	o.Emission = reflcol.rgb * _ReflectColor.rgb;
 	o.Alpha = reflcol.a * _ReflectColor.a;
