+++ DefaultResources/Reflect-Glossy.shader
@@ -17,9 +17,9 @@ CGPROGRAM
 sampler2D _MainTex;
 samplerCUBE _Cube;
 
-float4 _Color;
-float4 _ReflectColor;
-float _Shininess;
+fixed4 _Color;
+fixed4 _ReflectColor;
+half _Shininess;
 
 struct Input {
 	float2 uv_MainTex;
@@ -27,13 +27,13 @@ struct Input {
 };
 
 void surf (Input IN, inout SurfaceOutput o) {
-	half4 tex = tex2D(_MainTex, IN.uv_MainTex);
-	half4 c = tex * _Color;
+	fixed4 tex = tex2D(_MainTex, IN.uv_MainTex);
+	fixed4 c = tex * _Color;
 	o.Albedo = c.rgb;
 	o.Gloss = tex.a;
 	o.Specular = _Shininess;
 	
-	half4 reflcol = texCUBE (_Cube, IN.worldRefl);
+	fixed4 reflcol = texCUBE (_Cube, IN.worldRefl);
 	reflcol *= tex.a;
 	o.Emission = reflcol.rgb * _ReflectColor.rgb;
 	o.Alpha = reflcol.a * _ReflectColor.a;
