+++ DefaultResources/Reflect-Parallax.shader
@@ -21,8 +21,8 @@ sampler2D _BumpMap;
 samplerCUBE _Cube;
 sampler2D _ParallaxMap;
 
-float4 _Color;
-float4 _ReflectColor;
+fixed4 _Color;
+fixed4 _ReflectColor;
 float _Parallax;
 
 struct Input {
@@ -39,14 +39,14 @@ void surf (Input IN, inout SurfaceOutput o) {
 	IN.uv_MainTex += offset;
 	IN.uv_BumpMap += offset;
 	
-	half4 tex = tex2D(_MainTex, IN.uv_MainTex);
-	half4 c = tex * _Color;
+	fixed4 tex = tex2D(_MainTex, IN.uv_MainTex);
+	fixed4 c = tex * _Color;
 	o.Albedo = c.rgb;
 	
 	o.Normal = UnpackNormal(tex2D(_BumpMap, IN.uv_BumpMap));
 	
 	float3 worldRefl = WorldReflectionVector (IN, o.Normal);
-	half4 reflcol = texCUBE (_Cube, worldRefl);
+	fixed4 reflcol = texCUBE (_Cube, worldRefl);
 	reflcol *= tex.a;
 	o.Emission = reflcol.rgb * _ReflectColor.rgb;
 	o.Alpha = reflcol.a * _ReflectColor.a;
