+++ DefaultResources/Reflect-ParallaxSpec.shader
@@ -23,9 +23,9 @@ sampler2D _BumpMap;
 samplerCUBE _Cube;
 sampler2D _ParallaxMap;
 
-float4 _Color;
-float4 _ReflectColor;
-float _Shininess;
+fixed4 _Color;
+fixed4 _ReflectColor;
+half _Shininess;
 float _Parallax;
 
 struct Input {
@@ -42,7 +42,7 @@ void surf (Input IN, inout SurfaceOutput o) {
 	IN.uv_MainTex += offset;
 	IN.uv_BumpMap += offset;
 	
-	half4 tex = tex2D(_MainTex, IN.uv_MainTex);
+	fixed4 tex = tex2D(_MainTex, IN.uv_MainTex);
 	o.Albedo = tex.rgb * _Color.rgb;
 	o.Gloss = tex.a;
 	o.Specular = _Shininess;
@@ -50,7 +50,7 @@ void surf (Input IN, inout SurfaceOutput o) {
 	o.Normal = UnpackNormal(tex2D(_BumpMap, IN.uv_BumpMap));
 	
 	float3 worldRefl = WorldReflectionVector (IN, o.Normal);
-	half4 reflcol = texCUBE (_Cube, worldRefl);
+	fixed4 reflcol = texCUBE (_Cube, worldRefl);
 	reflcol *= tex.a;
 	o.Emission = reflcol.rgb * _ReflectColor.rgb;
 	o.Alpha = reflcol.a * _ReflectColor.a;
