+++ DefaultResources/TerrainShaders/Details/VertexLit.shader
@@ -13,11 +13,11 @@ sampler2D _MainTex;
 
 struct Input {
 	float2 uv_MainTex;
-	float4 color : COLOR;
+	fixed4 color : COLOR;
 };
 
 void surf (Input IN, inout SurfaceOutput o) {
-	half4 c = tex2D(_MainTex, IN.uv_MainTex) * IN.color;
+	fixed4 c = tex2D(_MainTex, IN.uv_MainTex) * IN.color;
 	o.Albedo = c.rgb;
 	o.Alpha = c.a;
 }
